LCEL_APP/requirements.txt
langchain
langchain-openai
langchain-core
python-dotenv
langsmith
ipykernel
streamlit